17/11/2022 avances

+++ app/src/main/java/com/codekolih/producciontablet/clases/Bobinas.java
@@ -19,19 +19,19 @@ public class Bobinas implements Serializable {
 
         // Getter Methods
 
-        public float getBobinaId() {
+        public int getBobinaId() {
             return BobinaId;
         }
 
-        public float getTareaId() {
+        public int getTareaId() {
             return TareaId;
         }
 
-        public float getProduccionId() {
+        public int getProduccionId() {
             return ProduccionId;
         }
 
-        public float getProveedorId() {
+        public int getProveedorId() {
             return ProveedorId;
         }
